Fix import for running app as script
